Cleaning up 'test_model.py'
Made a few changes, primarily to fix the cathode/separator boundary.  I 
believe it works correctly.

Implementation is currently held up by having no derivative function for 
the separator - it acts as an infinite reservoir to contribute species 
to the cathode.

Next step - treat the entire separator as a single CV and implement 
conservation equations.