Wrapping up separator model.
Something in the separator model is still not working correctly.  I 
think it is time to table this, for the time being, and focus on the 
cathode model.

Boundary condition on cathode is not that there is a flux of Li+ ions at 
the cathode/separator boundary, equivalent to i_ext.