Added MHC, BV, and Marcus options to cathode rxns.
The cathode now has the option to switch between the
three kinetics forms independently of the anode. Also,
the cathode is now tuned to LFP getting 3.2 V at 0% SOC
and 3.3 V at 50% SOC (This seems to be in agreement with
the literature). Tried adding in SEI resistance, running
into issues with that though.