Add simjulations/cyclic_voltammetry.py
Adds a simulation type for cyclic voltammetry experiments.
This requires changing potentiostatic boundary conditions so that a
list of times and potentials are input, with linear interpolation used
to look up potential as a function of time in the electrode `residual`
function.
Note that this also, therefore, requires the time be provided as an
input argument to all electrode residual functions.